golang: update to build against tip using vanilla go tools
Mostly this commit just imports falcore as github.com/ngmoco/falcore.
The addional changes are to do with the current signals package. In
tip it's deprecated to exp/signal and will likely be removed for Go1
so it looks like any hot-restart code will have to be refactored
around runtime.Sigrecv() until a new (if any) signals package is
introduced.